backend/requirements.txt
fastapi
uvicorn
python-jose
passlib
sqlalchemy
pydantic_settings
asyncpg
pydantic[email]
python-multipart